+++ config/compiletime/apps/navkit_sim/StationaryGnss.hpp
@@ -4,8 +4,8 @@
 #pragma once
 
 #include "navkit/StationaryGnss.hpp"
+#include "navkit/app_support/EmulatorBinding.hpp"
 #include "navkit/app_support/GnssEmulator.hpp"
-#include "navkit/app_support/SensorId.hpp"
 #include "navkit/app_support/SimulationApp.hpp"
 
 #include <tuple>